+++ mods/conflict.go
@@ -5,7 +5,7 @@ import (
 	"github.com/stephenafamo/bob/clause"
 )
 
-type Conflict[Q interface{ SetConflict(clause.Conflict) }] func() clause.Conflict
+type Conflict[Q interface{ SetConflict(bob.Expression) }] func() clause.ConflictClause
 
 func (s Conflict[Q]) Apply(q Q) {
 	q.SetConflict(s())
@@ -15,7 +15,7 @@ func (c Conflict[Q]) Where(where ...any) Conflict[Q] {
 	conflict := c()
 	conflict.Target.Where = append(conflict.Target.Where, where...)
 
-	return Conflict[Q](func() clause.Conflict {
+	return Conflict[Q](func() clause.ConflictClause {
 		return conflict
 	})
 }
@@ -24,12 +24,12 @@ func (c Conflict[Q]) DoNothing() bob.Mod[Q] {
 	conflict := c()
 	conflict.Do = "NOTHING"
 
-	return Conflict[Q](func() clause.Conflict {
+	return Conflict[Q](func() clause.ConflictClause {
 		return conflict
 	})
 }
 
-func (c Conflict[Q]) DoUpdate(sets ...bob.Mod[*clause.Conflict]) bob.Mod[Q] {
+func (c Conflict[Q]) DoUpdate(sets ...bob.Mod[*clause.ConflictClause]) bob.Mod[Q] {
 	conflict := c()
 	conflict.Do = "UPDATE"
 
@@ -37,7 +37,7 @@ func (c Conflict[Q]) DoUpdate(sets ...bob.Mod[*clause.Conflict]) bob.Mod[Q] {
 		set.Apply(&conflict)
 	}
 
-	return Conflict[Q](func() clause.Conflict {
+	return Conflict[Q](func() clause.ConflictClause {
 		return conflict
 	})
 }