+++ mods/conflict_test.go
@@ -2,7 +2,6 @@ package mods
 
 import (
 	"github.com/stephenafamo/bob"
-	"github.com/stephenafamo/bob/clause"
 )
 
-var _ bob.Mod[interface{ SetConflict(clause.Conflict) }] = Conflict[interface{ SetConflict(clause.Conflict) }](nil)
+var _ bob.Mod[interface{ SetConflict(bob.Expression) }] = Conflict[interface{ SetConflict(bob.Expression) }](nil)